backend/requirements.txt
praw==7.7.1
python-dotenv==1.0.0
kafka-python==2.0.2
pyspark==3.5.0
py4j==0.10.9.7